Make PostgresDB obj encapsulate a conn from the pool
This isn't precise yet but the goal here is that operations on a
PostgresDB object take place on one connection and would encompass
transaction and that it is up to the client to commit or rollback at the
end.

This is fairly in line with how it's been working but now it is
pulling its connection from the pool and not allowing access to internal
private cursor/conn state.

The functions that were making a point of using a separate cursor i've
changed to now be classmethods that use a different connection from the
pool. This helps preserve the idea that an instance of this object
represents on logical transaction going on while still allowing this
object to be a common point of shared SQL/db operations in this system

re #1