Supply missing `portal_url` value
This was referencing it as `self.portal_url` but this wasn't in method,
or self from antywhere else.

Reading through the code around it, it looks like `portal_url` is in the
publisher row that is currently being processed; so pull the data from
there and pass as a parameter.