Split `idigbio_ingestion.update_publisher_recordset.update_db_from_rss`
Trying to make this a more manageable hunk I split the big function into
a couple smaller ones.

The original function is still the driver loop and does the error
handling and now there is a separate function that was the body of the
loop and processes one rss feed.

This shouldn't actually change any logic, all the same pieces just in
slightly different places and indentation levels.