Expand README requirements info, add test-requirements
Be more verbose in the README about system packages that are necessary
to run the ingestion as well as the api.

Include an `idigbio_ingestion/requirements.txt` to document the python
libraries necessary for that code as well as a `test-requirements.txt`
to indicate the python libraries necessary for testing.