First pass at media upload that can add a reference
for https://github.com/iDigBio/idigbio-python-client/issues/7 we want to
be able to add a new `filereference`(`url` in the media table) to an
existing etag.

With this we have the possibility of, during an upload every combination
of existing and non-existing etag and filereference.

This pass tries to handle that with a MediaObject class performing all
of the operations; the goal being to use this class in mediaing.py at
some point later