Test setup for MediaObjects
This is a bunch of tests that help verify MediaObject does its job
properly.

In order to get this to work I made a framework for creating a local
postgres idigbio database and dropping it each time a round.

In order for this to work you need to have:

 * A local postgres instance
 * the user `test`/`test` that can create databases

re https://github.com/iDigBio/idigbio-python-client/issues/7