Fix insert media bug in update_publisher_recordset
Previously it would only update the media record, it wouldn't insert the
media if it was new. This fixes that by adding a `ensure_media` method
to the `MediaObject` and uses it.

All of the methods on there now have `ensure_...` varieties; that's
really how we want to deal with that table is: make sure the row exists
correctly.

re #7