Add missing 'return etag'
In the update_publisher_recordset script the `upload_recordset` function
is supposed to take care of ensuring the file in storage and the
media/objects tables are updated. it then *should* return the etag back
to `harvest_file` which is dealing with the recordsets and setting the
`file_harvest_date` and the `file_harvest_etag`. the return statement
was missing, so we were getting back None and hence the missing value in
the db.

re #8