Don't swallow exception in upload_recordset
this try-except was introduced b/c it was thought that there was an
error uploading in here. However this except would cause the exception
to be swallowed and the calling `harvest_file` function would think that
the upload was successful.

We want this function to hard error if it does, not just log and
swallow. Any exeption in the `harvest_file` segment will be caught in
the parent `harvest_all_file` loop which will rollback the entire
transaction.

re #8