Connect to Biodiversity socket server lazily
Previously it tried to connect to the biodiversity socket server *very*
eagerly: on module import. This triggered a bunch of warnings during
testing/development, and causes extra connections on the live server by
idb processes that never use the server.

I refactored the `Biodiversity` class to only make the socket connection
when it is needed.

If there is a failure reading or writing to the socket it will
disconnect and try to re-establish at next use (`None` in the `_sock`
slot). A failure to connect triggers giving up (`False` in the `_sock`
slot).

This behavior of retrying once on failing to send and then giving up if
failing again persists from the previous implementation; although I
think now it may retry in slightly more cases.

Exception catching has been tightened to the expected `socket.error`s
and the log messages will print the exceptions message, e.g.
`[Errno 111] Connection refused`