A basic CLI and setup.py
This is a basic CLI implementation for the `idb` package and some lib
functions for further toplevel scripts in the `idigbio_ingestion`
package.

The setup.py is relatively minimal, we aren't trying to specify the
requirements in here, that stays handled in `requirements.txt`. This
allows the packages to be pip installed though making `PYTHONPATH`
issues easier as well as taking advantage of click's setuptools
integration.

Some things we're going for here:

 * clibase defines a lot of common options, e.g. --env, --config
   --verbose that effect the environment we're running in.
 * clibase is designed to be reused in all CLI scripts produced by this
   package

re #12