Use shared IDigBioStorage object, maintain fewer intermediate results
I had the derivatives process crash because it was out of file handles;
instead of having separate IDigBioStorage instances for every object
that each has a connection pool cache use one.

I originally thought that it would be necessary to give each one a
separate instance because I wasn't sure if boto was green thread
safe. I've written a couple of other gevented proceses that used boto
and haven't had any problems with it.