Fetch files from storage atomically
If the process is interrupted (Ctrl-C) during download it will leave a
partial file on disk. This uses the atomicfile library to download to a
tempfile and then rename to destination on completion.

Add some methods on the `IDigBioStorage` object to use this, as well as
optionally doing an md5 check against the downloaded result. Right now
we're checking the md5 that was calculated by boto during download
against the etag/md5 we calculated when we stored the object
originally. the S3 interface presents back an etag as part of metadata
when fetching the file but that might not be an md5 hash if the file was
posted in multiple parts.