Turn gevent off completely
The monkeypatching was originally disabled because of concerns about
memory usage. In my testing with a low (100) maxsize bounding the result
pool i showed no memory difference between a gevent pool with
`monkey.patch_all` and just using `itertools.imap`.

However the imap was a few percent faster.

This testing was done without actually posting index to elastic search
which could notably change the result. Since we post to elasticsearch in
batches of 10000 the maxsize would probably need to be >5k (and probably
> 10k) in order to have any effect.