Don't import config until processing cmdline
If config is imported before click has a chance to read environment
variables out of it then config will set environment based on json
config files and then click will think those values are being provided
in the environment.

By delaying the import click reads the environment variables that are
actually specified before/with process launch and it can then use those
to override what is specified in the json config.

re #12