Enforce specified mime matches detected_mime
When creating a MediaObject from a file/object, if a mime type was
passed in then enforce that the detected type matches the passed in
type.

There are some buckets like debuglog and datasets where we aren't
limitting what is put in them; so to still have some sanity checks that
what was posted is valid this is a good baseline.