Add toplevel exception logger to CLI entrances
when an exception goes uncaught out of the progam it will print into the
console, but would not have been logged into the logfile.

This adds two decorators `fntimed` and `fnlogged` that can be put onto a
function to record the time it takes and log if it raises an error.

I've added `fnlogged` to all the cli entries such that if it crashes we
get it in the logfile.

re #12