Pipelined mediaing fetcher per domain
This is a rework of mediaing with a bunch of changes. The most
fundamental one is that we parallelize across domains.

The general process is:

1. Find all outstanding urls to fetch.
  * no longer ignores URLs with a bad or no status but with an etag
2. Group urls by prefix, prefix is the scheme+domain
3. For each prefix group launch a subprocess that will fetch and those
   urls and update the db.

This can be run as a one shot and it will run to completion or it can
be setup to run in a timed loop (managed by supervisord). The loop
logic is

1. run the one shot and keep track of what prefix-groups were started.
2. When the loop timer expires:
  * check if any prefix groups have finished
  * Build an ignore list of prefixes that are still running and ignore
    those when fetching the next list to work on

re #6: more unified code paths for media fetching, saving to the db and
validation

re #13: fetching posted urls on a reasonable schedule involves this type
of continuous rechecking