Rework media validation
This introduces a new validation dictionary: the list of allowed mimes
in each bucket.

With this we get better testing of the `datasets` bucket; previously we
allowed anything in there, now it must be either in
`{'text/csv','text/plain','application/zip'}`.

The validation is more consistent now, mediaing and media api use the
same code paths.