Use gipc in db_check to launch with clean environment
When using multiprocessing the fork doesn't leave things in a very good
state. https://gehrcke.de/gipc/ helps clean that up and provides a more
stable state.

We were getting an error in db_check (#15) that we think is related to
this leftover state after a fork. By using gipc to handler the fork I
think it will fix it.