Extend the timerange on updatedb --daily
I had set this previously to 25hrs so there would be some overlap; but
Dan pointed out that recordsets are ingested in a transaction and some
of these transactions can be rather long.

by extending this window to the previous 2 days checked every day that
should eliminate any missing there.