Merge two tests directories
The code that was in `idb/tests` had a lot of really good fixtures for
setting up a test DB to work with.

In order to be able to use this for tests in `idigbio_ingestion` i
converted to having one toplevel tests directory with subdirectories now
for the various pieces being tested.

re #3