Pluralize 'mediarecords' to conform to conversions api
Everywhere else in the conversions api the `type` (e.g. 'records',
'mediarecords') is plural.

Previously when updatdb had called the functions `get_accessuri` and
`get_media_type` those functions were ignoring the type parameter. In
9ff286541fc9c21fc496b834493fce318c4da835 it was changed to only return a
dictionary with keys if it was `type='mediarecords'` so that `grabAll`
wouldn't return extraneous fields.