Fix index_record incorretly passing parent on delete
During mediarecord deletion we were passing the `parent` which in the Postgres data
model is the recordset uuid. In the ElasticSearch data model the
`_parent` is the associated entry of type 'record'.

It couldn't delete the entry when the `_parent` was specified incorrectly
and the info that index_record was producing had already stripped out
that data; this fixes that.

Longer term it would probably be good to collapse index_record and
bulk_formatter into a single function. Originally they were split
because `index_record` was doing the indexing, but then it gained a flag
to just return a work item to pass to bulk. We long ago switched to only
using the bulk codepath but this duplication still causes some
confusion...

This helps fix https://www.idigbio.org/redmine/issues/1949