A basic readonly setup
This has a couple of points; when getting a connection from the DB it
can specifically opt for a readonly on not connection. This is not
intended to be used much but is rather useful for testing.

This helps assert that code paths that aren't supposed to modify
the DB don't and that lots of the code still works even when the DB is
set in read only mode.

re #3