Guarantee return value in tropicos fetcher
When it was running out o retries it was possible to fall out of the
bottom of the loop with the default return value of None. This causes
problems down the pipeline. This was revealed recently due to separate
bugfix in 800599a0b4a03e4206df8b69c5bd2ff3acefeb6d that enabled a
previously untrod codepath.

I think the loop logic is cleaned up to fix that problem as well as a
safeguard at the bottom so any future fall through would return self
anyways.