Move all requirements back into setup.py
Reading through https://caremad.io/2013/07/setup-vs-requirement/ and
similar I think this may be a better scheme. I've updated README and it
is certainly simpler, at least for the story of "I want everything
installed" which is the common one.

Bump to version 2 as we're breaking some past behaviors and links, still
a beta release while I test it out.