Handle deletes using bulk codepath in index check/resume
Previously it would handle deletes as individual calls at the end of the
loop, now it will process those through the same bulk indexing codepath
that everything else goes through.

This was motivated by the fact that the individual delete calls didn't
pass the `_parent` attr that ES wants, we fixed this bug for the bulk
codepath in 37328571085ebf1cfe72a0318484969762a02d51 but was still
missing here.