Verify etag on upload if present in POST
Version 0.8.1 of the `idigbio-python-client` started calculating the
etag on the client and including that in the post. This patch now
verifies it matches the server calculated version if it is presented by
the client.

We've seen some suspicious behavior from clients over wireless and as
these are large files it makes sens to verify the upload integrity when
we can.