Implement delete-by-query in bulk index pipeline
We thought we had this fixed previously but today we found it behaves
differently for UUIDs that are associated with records and aren't
associated with record

If we send the following parent/routing values on a `get` request (and
we think delete behaves the same way) we get very contradictory
behavior.

                False    'null'    'false'   'False'     0    '*'
------------------------------------------------------------------
UUID w/ rec   |   T         T        T          F        F     F
UUID w/o rec  |   F         F        F          T        T     T

This patch says: "If we know routing information, use it; if we don't
then query to find it and use that". This avoids trying to speciy any
magic wildcard value b/c the above table demonstrates that none exists
that will work for both cases.