Always query ES for routing info on delete
During reindexing (especially deletes), the sibling information in the
DB might not match what had previously been stored in elasticsearch.

If we're trying to delete a mediarecord it doesn't matter what sibling
info we have, the only thing that matters is what elasticsearch has and
making sure the delete gets processed successfully.