Get redis conn based on ENV; make it a memoized func
Previously it always pointed at the URL of the production redis DB. Now
it points at the same result DB configured for celery. This fixes a
potential bug where similar download request between beta and prod has
the result in one location but not in the other, but the code attempts
to fetch it in both places anyways.