Remove `use_string_io` branch from download
This is an unused path. It could be good as it eliminates really slow
disk IO but the cost of memory is too high.

Right now it is making the code path harder to read/reason about. If we
want to put it back in it would probably be better to handle that at a
wrapper function layer where we pass in the stream to write to.