Split generate_files into pieces
The main goal here is turn the file generation for dwc archives into a
generator. I.e. instead of generating a bunch of intermediate files and
then creating a zipfile out of them, generate a file and immediately add
it to the zip before continuing.

`generate_queries` func now takes `core_type` argument and will handle
requiring images if it is set to mediarecords, something that was in
generate_files before.