Return early if no records
Previously it would do a bunch of template formatting and write file to
disk and then if there were no records it wouldn't return the
filename (which also means that it would never be unlinked).

Now do the check for "are there any records" earlier and if we're going
to return None do so sooner, before writing the file to disk.