Move retry loop into core upload logic, rework large uploads
We implemented an upload retry loop a bit back in the mediaing fetcher,
this moves it into the core of the storage lib.

Previously the upload of large files would split the file into multiple
files on disk and then upload from those. We don't need to split the
file on disk, we can just pass it fd that has been moved to the correct
offset.