Prefetch rss contents and parse those
Previously we did a GET on the RSS url to verify the remote endpoint was
up, but then discarded the results and had feedparser fetch the url
again.

Now we save the results of the first fetch and have feedparser just
parse that string rather than refetching. This is because we had at
least one case where the initial connection succeeded but then we
found (4 days later) it was hung reading from the same url, we believe
in the feedparser section.