Don't use fontconfig; use `DejaVu Sans:bold`
Previously we brought in font config as a dependency to badly search for
the ubuntu font, which isn't (easily available) in the debian based
docker image we're building.

Instead of specifying fontconfig as a dependency we're now requiring the
system package `fonts-dejavu-core` as a dependency and depending on its
behavior.