Revert "fetcher: update media objects with the same transaction"
This reverts commit 9a2178c000f10659dd9d96640b636410f7e3ab6d.

I made the mistake in there of ignoring failures. If it fails then
"upload_to_storage" exits early and we wouldn't save this failure to the
DB. I think that still makes sense and we don't want to change
that. Instead restore the old behavior of updating the DB in a separate
block that applies to everything.