Move migrate (old media api) into separate module/timer invocation
This had been wrapped into derivatives, while that is somewhat relevant
it isn't exactly the same process. This was already encoded in that it was
running in a separate thread in the derivatives process.

There's really not any shared code between these two things though and
moving it into its own module makes it cleaner and then we don't get
threadpool/processpool conflicts between derivatives and migration