Still touch the DB even if object exists in ceph.
Previously if the object already existed in ceph we short circuited out
of the function and didn't create the media_object link (bug).

We found that there are image copies/multiple urls out there and now the
more recent recordset to mention these urls doesn't have any of the
media_object correct.

This fixes the bug in fetcher, we still need to figure out a data fix.