Use `IDigBioStorage.retry_loop` in derivatives
I had originally made the upload loop in the media fetcher, but then
found it to be useful so promoted it up to the storage
class. Derivatives should use it because it is useful.

Since now it is used more generally for upload/downloads/setmetadata
operation and it retries on any error boto is likely to throw it has
been made public and the name generalized as `retry_loop`