Script to go through recent derivatives and make set public/mime
We had some errors in the derivatives script recently where it could
upload the image but then not complete setting it public and mime.

This script goes through the objects that were modified in the last
couple months and checks up on them. It might also be a useful
template for other tasks that are mapping over ceph keys.