Fix reset-public script
This is using the copy-object with metadata replacement trick to
update the metadata and otherwise only putting the ACL. This has been
tested much better and is safe.

Otherwise it is checking for missing derivatives and updating the db
to trigger rederive.

It also checks for 0 byte objects or ones that don't match the etag
which could be ones that weren't fixed in the previous cleanup efforts