Integrate all repairs into new fixup-objects script
except for restore from backup.

This is meant to be a main script for going through objects from the
DB and ensuring they are in as correct a state as we can make them
without going to backups.

 * Check+Fix ACL/mime of objects that are present and correct
 * Copy from fullsize if applicable
 * Attempt redownload

TODO: test/debug  (I need to be on other computer to do this).