Loosen test on the dqs field
This test wasn't actually verifying any logic, just that the number
calcuated now matched a number calculated at some point previously. In
the meantime that algorithm has shifted a bit and teh data has changed
too.

Instead do a more sane assert that a number is produced and bounded as
expected. This is really all that we're in a position to guarantee at
the moment.