Run fixup-objects scripts on derivative-less objects
When we originally constructed this query in the fixup scripts it was
intended to catch bad permissions on derivative objects so it only
made sense to look at objects that had derivatives.

Since then we've expanded it to checking for a host of other
faults/errors that would make sense to check on the original object too.