Download task launching refactor
This changes the download api endpoint to:

Store data in redis prefixed with `downloader:`, so we can easily
distinguish this set of uuids from other sets of uuids

Store data about the download (query params, created date) associated
with the task uuid and have the query hash point at that. We can keep
task metadata around for a while to answer /status/ queries about it
but we want the query hash that points to that task to expire
relatively quickly (about 1 day) so that new download requests for the
same query get new data.

Fix the race condition of a second person asking for the same download
while the first one is still being generated would have been dropped.

When checking on the status of a failed download remove the link from
query hash to task such that new download requests for the same query
will cause a retry.