Trying to setup travis builds (WIP)
This includes the .travis.yml similar to how I did it on
idigbio-search-api: it will attempt to build the docker container and
run the tests in there, and then if it passes uploads that to
dockerhub.

I've included the `test` option in the docker build, left out
originally to save space, but it only adds a trivial amount of space
and makes the testing much easier (we can test the image directly
instead of making a new one based on it that has the testing code and
then test that).

I have uploading disabled for now until we get this sorted out (given
that it is setup for after_success this is a bit redundant).

Enable the postgresql addon for travis and try and get it setup. We
might also pull a couple more docker containers to provide a fake s3
interface too.