Make healthz celery task a bit more verbose
I'm trying to include the node name that actually executes the task
but I can't find anyway to introspect that. Instead I'm going to set
it on the cmdline based on environment variable and then try to read
that here, falling back to the hostname. As this i will be running in
a docker container the hostname isn't terribly useful though.