use the value in getty thesaurus for nation names
For clarification, we choose to use the "English" nation name rather than the "preferred" nation name.

In Getty, the "preferred" name is the name most commonly used in the vernacular (local) language,
it is not the most common name used by the primary iDigBio user base (english speakers in the United States).

Some examples.

Getty English, Getty Preferred, ISO Short, ISO Long
"United States", "United States", "United States of America (the)", "the United States of America"
"South Korea", "Han'guk", "Korea (the Republic of)", "the Republic of Korea"
"Democratic Republic of the Congo ", "Congo, République Démocratique du", "Congo (the)", "the Republic of the Congo"
"Congo", "Congo (Republic of)", "Congo (the)", "the Republic of the Congo"