Follow Nathan's practice of naming one-off scripts with hyphens
Naming the script with a hyphen prevents it from being imported as
a module into other programs.

Possibly encourages the code to be incorporated into the actual
codebase if it becomes apparent that it is used more than once.

The downside is that the functions already written that might be
useful in other one-off scripts cannot be imported, potentially
leading to more copy-and-paste code instead.

Nathan said that it also makes it more difficult to develop
parts interactively via ipython, etc.