create new script analyze-all-eml.py for parsing all eml files
This script gets list of all dataset metadata file links from the
database where ingest=true, downloads the eml files from the sources,
and parses out the data rights (license) strings.

Results are saved to an output file.

This script is useful to compare the effects of eml parsing changes
across all of iDigBio datasets (run it before the changes and
after the changes, use diff to compare output files).