update eml parsing to properly detect known licenses
I noticed that a large number of licenses were not being detected properly where I could
see a reasonable license string in the various permutations of the intellectualRights
section.

This commit makes a number of changes:

1. Hard-code the pyquery parser to 'xml'.

By default, pyquery changes the parser it uses (and thus behavior) depending on certain
attributes of the data it finds.

For example, case-sensitive XML tags such as "intellectualRights" caused the parser to
switch to 'html'.  This caused no child elements to be found under "intellectualRights"
even though they existed in the document.  See this:

https://github.com/gawel/pyquery/issues/84

2. Updated rights strings data tables to include additional license strings that
can be mapped to known licenses.

3. Other improvements to the eml.py including remove extraneous comments.

4. Added more sample files to tests data and activated all of the eml tests in
test_eml.py since the eml parsing now works properly for all of the test cases.