Improve logging in failure cases during ingestion.
Provide a generic exception handler in the delimited file code to catch things like NULL byte exceptions.
Change the unhandled exception behavior of process file so that it always cleans up and returns metadata
to improve the ability of ds_sum_counts to flag files with fatal exceptions as suspect.