stop logging tracebacks when StopIteration is reached
The iterator raises StopIteration when read.next() runs out of rows.

We were previously noticing this in a generic Exception catch which
caused the traceback to be logged.

Since "we reached the end of file" is a known scenario, we now
catch this Exception cleanly and re-raise it to stop reading
lines from input in the calling code.

Continue to log a debug line but do not include the traceback.