catch NUL Bytes and other csv reader errors
We now do not abort the entire file on an exception caused by a
NUL Byte (python library mistakenly reports "NULL" byte).

It seems that our implementation with csv reader does a good job
of throwing away the offending line and continuing onward with
the next line. Alex warns that there is danger to continue
processing the file. A nul byte is clear indication of bad
encoding in the file or other problem, it is possible that the
position in the bytestream will be lost.  If the end of line
is lost after a nul byte is encountered, then most likely one
or more (or all) of the remaining lines in the file will be
"wrong".

The old verbose logs aborted processing with OtherException.

2017-05-05 18:55:02.870 WARNI idb.699a6c55-fded-431b-8657-a0239b21e147.occurrence? OtherException line contains NULL byte: 699a6c55-fded-431b-8657-a0239b21e147_extracted/occurrence.txt Line 1209 (92)
2017-05-05 18:55:02.870 INFO  idb.699a6c55-fded-431b-8657-a0239b21e147.occurrence? Traceback (most recent call last):
  File "/root/idb-backend/idigbio_ingestion/lib/delimited.py", line 129, in readline
    lineArr = self._reader.next()
  File "/usr/local/lib/python2.7/dist-packages/unicodecsv/py2.py", line 117, in next
    row = self.reader.next()
Error: line contains NULL byte

The new code logs the csv error and continues onward to the
next row in the file:

09:59:14.402 ERROR idb.fabcdc12-9d29-4bd2-912b-176e71818144.occurrence჻ csv ERROR at line 19699, possibly a NUL Byte or encoding mismatch
09:59:14.403 DEBUG idb.fabcdc12-9d29-4bd2-912b-176e71818144.occurrence჻ Traceback (most recent call last):
  File "/home/dstoner/git/idb-backend/idigbio_ingestion/lib/delimited.py", line 129, in readline
    lineArr = self._reader.next()
  File "/home/dstoner/git/Envs/idb-backend/local/lib/python2.7/site-packages/unicodecsv/py2.py", line 117, in next
    row = self.reader.next()
Error: line contains NULL byte

10:00:22.520 ERROR idb.fabcdc12-9d29-4bd2-912b-176e71818144.occurrence჻ csv ERROR at line 48908, possibly a NUL Byte or encoding mismatch
10:00:22.521 DEBUG idb.fabcdc12-9d29-4bd2-912b-176e71818144.occurrence჻ Traceback (most recent call last):
  File "/home/dstoner/git/idb-backend/idigbio_ingestion/lib/delimited.py", line 129, in readline
    lineArr = self._reader.next()
  File "/home/dstoner/git/Envs/idb-backend/local/lib/python2.7/site-packages/unicodecsv/py2.py", line 117, in next
    row = self.reader.next()
Error: line contains NULL byte

And even better, our normal execution logging style (single -v)
gives single error logline:

$ idigbio-ingestion -v db-check fabcdc12-9d29-4bd2-912b-176e71818144
10:00:31.635 INFO  idb.db-check.fabcdc12-9d29-4bd2-912b-176e71818144჻ Starting db_check ingest: False
10:00:31.635 INFO  idb.db-check.fabcdc12-9d29-4bd2-912b-176e71818144჻ Processing fabcdc12-9d29-4bd2-912b-176e71818144, type: application/zip
10:01:27.049 ERROR idb.fabcdc12-9d29-4bd2-912b-176e71818144.occurrence჻ csv ERROR at line 19699, possibly a NUL Byte or encoding mismatch
10:02:35.205 ERROR idb.fabcdc12-9d29-4bd2-912b-176e71818144.occurrence჻ csv ERROR at line 48908, possibly a NUL Byte or encoding mismatch
10:07:34.351 ERROR idb.fabcdc12-9d29-4bd2-912b-176e71818144.occurrence჻ csv ERROR at line 182308, possibly a NUL Byte or encoding mismatch