add ability to pause ingestion on a recordset without changing its ingest flag
We now have two toggles in the database.

ingest is the flag that says "iDigBio would like to ingest this dataset".

ingest_is_paused is the flag that says "something is going on with the
currently published dataset file, please do not ingest right now while
we investigate or wait for a re-publishing event from the provider".

It is possible that in the future we can leverage the new paused flag
for more automatic ingestion.  We could automatically pause ingestion
on any recordsets that appear in the suspects file, for example.