refactor to move arg checking into a function, globals handling, reporting
The keyword "global" inside a function means we want to be able to
change the value outside the function scope.

Modified reporting to display relevant summary depending on which
flags / operations were used.

This version does not yet log the list of deleted objects to a
file or the database.