add delete object capability, still missing delete status tracking
I had some confusion about the '--reverify' argument. I thought
it was a boolean but it appears the original script used this as
a limiter to only reverify specific statuses. This means the status
must be known ahead of time.

Added 'ANY' reverify option to allow reverify regardless of previous
status.

Ceph objects can now be deleted via '--stash-and-delete'.

Have not yet implemented tracking in ceph_objects for items that
have been deleted. That is important and will come next.