docker-compose container naming and improve tacklebox startup loglines
Originally I thought tacklebox would be better to start "if needed"
from the host.

However, this makes it a little less usable since one has to first
locate the proper docker network and start tacklebox on that
particular network.

For now, letting docker-compose start tacklebox so it is already on
the same network.

The containers on the same docker network can all reach each other
by their short names thanks to docker networking.

  dev-database
  dev-storage
  dev-elasticsearch