Revert "use proper pyproj Transformer pattern"
This reverts commit 3b9db4529f37da82fe8f76d393a0d09b9b430a50.

We do not have good test coverage around these transforms from one
datum to another and a lot geo-domain knowledge is needed to write those
tests.

Reverting this for now, since the pyproj functions we were using
are deprecated, not yet removed, and not needed to complete the
python3 conversion.