feat: add Redis authentication capability
Specify authentication password by running this application with
set environment variable 'IDB_REDIS_AUTH'
or by setting $.env.IDB_REDIS_AUTH in idigbio.json.

Note: Unlike idb-search and idb-portal (Node.js),
this application (Python) considers supplying a password to a Redis
database that does not need one to be an error, instead of a warning.
I have spent quite a bit of time trying to de-escalate this occurrence
to a warning, but it seems a lot of application configuration is
"set in stone" the moment you create a Celery instance;
setting Celery.conf.broker_url afterward
(even before any tasks are run) will show tasks throwing errors due to
using outdated settings.

With get_redis_connection_params() modified to remove the password from
broker_url when getting the specific error above, I tried:
- Adding a consumer bootstep to reset broker_url::
  ```
  class InConsumerConnectionStep(bootsteps.StartStopStep):
      requires = {'celery.worker.consumer.connection:Connection'}

      def start(self, consumer):
          _ = get_redis_connection_params()
  app.steps['consumer'].add(InConsumerConnectionStep)
  ```
  Result: Tasks still using old broker_url

- Restarting the broker connection::
  ```
  from celery.worker.control import Panel
  @Panel.register
  def reset_celery_broker_connection(state):
      state.consumer.connection.close()
      return {'ok': 'connection reset'}
  app.control.broadcast('reset_celery_broker_connection')
  ```
  Result: Tasks are received(?), but logging stops
    (loss of monitoring capability)

- Handling failed tasks and retrying::
  ```
  @signals.task_failure.connect
  def handle_celery_task_failure(sender, exception, **kwargs):
      import redis
      if isinstance(exception, redis.ResponseError) \
              and exception.args[0] == 'Client sent AUTH, but no password is set':
          _ = get_redis_connection_params() # trigger broker_url correction
          app.control.broadcast('reset_celery_broker_connection')
          sender.retry(countdown=1, throw=False)
  ```
  Result: Task infinite loops with failure

If you want to give it a shot, be my guest.
After putting in over a week on this,
I think we'll just leave it as a fatal error.